requirements.txt
aioredis==1.3.1
appdirs==1.4.4
APScheduler==3.7.0
apt-clone==0.2.1
apturl==0.5.2
arabic-reshaper==2.1.3
argon2-cffi==20.1.0
asgiref==3.7.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.8.2
black==21.7b0
bleach==3.3.0
blinker==1.4
Brlapi==0.7.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
click==8.0.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crypto==1.4.1
cryptography==3.4.7
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.16
decorator==5.0.5
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.2
distro==1.5.0
Django==4.2.7
django-apscheduler==0.5.2
django-cors-headers==4.3.1
django-extensions==3.1.2
django-filter==2.4.0
djangorestframework==3.12.4
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.1
entrypoints==0.3
et-xmlfile==1.1.0
extension==1.0.0
filelock==3.0.12
Flask==2.0.1
future==0.18.2
gevent==22.10.2
greenlet==2.0.2
grpcio==1.16.1
hiredis==2.0.0
html5lib==1.1
httplib2==0.14.0
hyperlink==21.0.0
idna==2.10
ifaddr==0.1.6
IMDbPY==6.8
importlib-resources==5.12.0
incremental==21.3.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==2.11.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==6.1.13
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==1.0.2
mypy-extensions==0.4.3
Naked==0.1.31
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nemo-emblems==5.2.0
nest-asyncio==1.5.1
netaddr==0.7.19
netifaces==0.10.4
nose==1.3.7
notebook==6.3.0
oauthlib==3.1.0
onboard==1.4.1
opcua==0.98.13
openpyxl==3.0.7
packaging==20.9
PAM==0.4.2
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.8.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyparted==3.11.2
PyPDF2==1.26.0
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.1+ubuntu0.20.4.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.18.0
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pytz==2021.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.8.3
reportlab==3.5.34
requests==2.25.1
requests-file==1.4.3
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==21.1.0
setproctitle==1.1.10
shellescape==3.8.1
simplejson==3.16.0
six==1.15.0
soupsieve==1.9.5
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
terminado==0.9.4
testpath==0.4.4
texttable==0.9.0
tinycss2==1.0.2
tldextract==2.2.1
tomli==1.2.1
tornado==6.1
traitlets==5.0.5
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.8.0
tzlocal==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.26.4
virtualenv==20.4.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.32.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xhtml2pdf==0.2.5
xkit==0.0.0
xlrd==1.1.0
youtube-dl==2021.4.26
zeroconf==0.24.4
zipp==3.15.0
zope.event==4.6
zope.interface==5.4.0
